amended the undo functionality for delete. Now calls a separate 'restore' endpoint

+++ ClientApp/src/Pages/AdminSightingsList/AdminSightingsList.tsx
@@ -5,24 +5,32 @@ import {
   fetchUnconfirmedSightings,
   ListSightings,
   Sighting,
-  deleteSighting
+  deleteSighting,
+  restoreSighting
 } from "../../Api/apiClient";
 import "./AdminSightingsList.scss";
 
 export function TableRow(data: Sighting): JSX.Element {
 
   const [confirmClicked, setConfirmClicked] = useState(false);
+  const [deleteClicked, setDeleteClicked] = useState(false);
   
   function ConfirmSightingRequest(id: number) {
     confirmSighting(id)
     .then(() => setConfirmClicked(!confirmClicked));
-  }
-
-  const [deleteClicked, setDeleteClicked] = useState(false);
+  }  
   
   function DeleteSightingRequest(id: number) {
-    deleteSighting(id)
-    .then(() => setDeleteClicked(!deleteClicked));
+    if (!deleteClicked) 
+    {
+      deleteSighting(id)
+      .then(() => setDeleteClicked(!deleteClicked));
+    } else 
+    {
+      restoreSighting(id)
+      .then(() => setDeleteClicked(!deleteClicked));
+    }
+    
   }
   
   return (
